requirements.txt
streamlit
pandas
pyzbar
pydub
Pillow
